move all the settings to newly created ComposeSettings class
